pyproject.toml
[tool.poetry]
name = "aruruu-bot"
version = "0.1.0"
description = ""
authors = ["domsupamoe"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.8"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
